pyproject.toml
[project]
name = "claude-db-tools"
version = "1.0.0"
description = "MCP Server for PostgreSQL access - Tools for Claude Code to interact with Cloud SQL databases"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Garantis Team"}
]
keywords = ["mcp", "claude", "postgresql", "database", "tools"]

dependencies = [
    "mcp[cli]>=1.0.0",
    "psycopg2-binary>=2.9.9",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
claude-db-tools = "src.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
